requirements.txt
mock
pycrypto
